requirements.txt
fastapi
uvicorn[standard]
numpy
scikit-learn
pydantic
pytest
pytest-asyncio
httpx
pyjwt-2.12.0

asyncpg
psycopg2-binary
yandex-pgmigrate

mlflow

aiokafka

prometheus-client
